created an edit form for issues
