+++ app/issues/new/loading.tsx
@@ -1,11 +1,10 @@
-import React from "react";
 import Skeleton from "react-loading-skeleton";
 
 const NewIssueLoading = () => {
   return (
     <div className="max-w-xl">
       <Skeleton />
-      <Skeleton height='20rem'/>
+      <Skeleton height="20rem" />
     </div>
   );
 };
